fix: method clearAppCache on iOS

+++ ios/ClearCache.swift
@@ -6,7 +6,8 @@ class ClearCache: NSObject {
         let cachPath: String = NSSearchPathForDirectoriesInDomains(.cachesDirectory, .userDomainMask, true)[0]
         let attrs = try? FileManager().attributesOfItem(atPath: cachPath)
         if (attrs != nil) {
-            resolve(attrs?[.size] ?? 0)
+            let size: Int = attrs?[.size] as! Int
+            resolve(size)
         } else {
             resolve(0)
         }
@@ -15,14 +16,10 @@ class ClearCache: NSObject {
     @objc(clearAppCache:withRejecter:)
     func clearAppCache(resolve:RCTPromiseResolveBlock,reject:RCTPromiseRejectBlock) -> Void {
         let cachPath: String = NSSearchPathForDirectoriesInDomains(.cachesDirectory, .userDomainMask, true)[0]
-        let files: [String]? = FileManager().subpaths(atPath: cachPath)
-        if (files != nil)  {
-            for p in files! {
-                let path = cachPath + p
-                if (FileManager().fileExists(atPath: path)) {
-                    try? FileManager().removeItem(atPath: path)
-                }
-            }
+        let files: [String?] = FileManager().subpaths(atPath: cachPath) ?? []
+        if (files.count > 0 && files[0] != nil)    {
+            let path = cachPath + "/" + files[0]!
+            try? FileManager().removeItem(atPath: path)
             resolve(true)
         } else {
             reject("error", "files is nil", false as? Error)
